+++ Restaurant/Restaurant/View/kitchen/uc_spendingRequest.xaml.cs
@@ -75,7 +75,7 @@ namespace Restaurant.View.kitchen
 
         static private int _SequenceNum = 0;
         static private int _Count = 0;
-        static public int _invoiceId;
+        static public long _invoiceId;
         #region barcode
         DateTime _lastKeystroke = new DateTime(0);
         static private string _BarcodeStr = "";
@@ -456,7 +456,7 @@ namespace Restaurant.View.kitchen
                         // get item matches the barcode
                         if (unit1 != null)
                         {
-                            int itemId = (int)unit1.itemId;
+                            long itemId = (int)unit1.itemId;
                             if (unit1.itemId != 0)
                             {
                                 int index = billDetails.IndexOf(billDetails.Where(p => p.itemUnitId == unit1.itemUnitId && p.OrderId == 0).FirstOrDefault());
@@ -632,7 +632,7 @@ namespace Restaurant.View.kitchen
                 invoiceItems.Add(itemT);
             }
             // save invoice in DB
-            int invoiceId = await FillCombo.invoice.saveInvoiceWithItems(invoice,invoiceItems);
+            long invoiceId = await FillCombo.invoice.saveInvoiceWithItems(invoice,invoiceItems);
             invoice.invoiceId = invoiceId;
             if (invoiceId != 0)
             {
@@ -1139,7 +1139,7 @@ namespace Restaurant.View.kitchen
                 {
                     for (int i = 0; i < w.selectedItems.Count; i++)
                     {
-                        int itemId = w.selectedItems[i];
+                        long itemId = w.selectedItems[i];
                         await ChangeItemIdEvent(itemId);
                     }
                     refrishBillDetails();
@@ -1439,7 +1439,7 @@ namespace Restaurant.View.kitchen
         }
         #endregion
         #region billdetails
-        private void addRowToBill(string itemName, int itemId, string unitName, int itemUnitId, int count)
+        private void addRowToBill(string itemName, long itemId, string unitName, long itemUnitId, int count)
         {
             // increase sequence for each read
             _SequenceNum++;
@@ -1481,7 +1481,7 @@ namespace Restaurant.View.kitchen
         //    DataGrid_CollectionChanged(dg_billDetails, null);
 
         //}
-        public async Task ChangeItemIdEvent(int itemId)
+        public async Task ChangeItemIdEvent(long itemId)
         {
              item = FillCombo.purchaseItems.ToList().Find(c => c.itemId == itemId);
