+++ Restaurant/Restaurant/View/reports/purchaseReports/uc_invoicePurchaseReports.xaml.cs
@@ -94,10 +94,10 @@ namespace Restaurant.View.reports.purchaseReports
 
         /*************************/
 
-        ObservableCollection<int> selectedBranchId = new ObservableCollection<int>();
-        ObservableCollection<int> selectedPosId = new ObservableCollection<int>();
-        ObservableCollection<int> selectedVendorsId = new ObservableCollection<int>();
-        ObservableCollection<int> selectedUserId = new ObservableCollection<int>();
+        ObservableCollection<long> selectedBranchId = new ObservableCollection<long>();
+        ObservableCollection<long> selectedPosId = new ObservableCollection<long>();
+        ObservableCollection<long> selectedVendorsId = new ObservableCollection<long>();
+        ObservableCollection<long> selectedUserId = new ObservableCollection<long>();
 
 
         private async void UserControl_Loaded(object sender, RoutedEventArgs e)
@@ -279,7 +279,7 @@ namespace Restaurant.View.reports.purchaseReports
             return xx;
         }
 
-        private void fillPieChart(ComboBox comboBox, ObservableCollection<int> stackedButton)
+        private void fillPieChart(ComboBox comboBox, ObservableCollection<long> stackedButton)
         {
             List<string> titles = new List<string>();
             IEnumerable<int> x = null;
@@ -338,7 +338,7 @@ namespace Restaurant.View.reports.purchaseReports
         }
 
 
-        private void fillColumnChart(ComboBox comboBox, ObservableCollection<int> stackedButton)
+        private void fillColumnChart(ComboBox comboBox, ObservableCollection<long> stackedButton)
         {
             axcolumn.Labels = new List<string>();
             List<string> names = new List<string>();
@@ -471,7 +471,7 @@ namespace Restaurant.View.reports.purchaseReports
             DataContext = this;
             cartesianChart.Series = columnChartData;
         }
-        private List<ItemTransferInvoice> fillPdfList(ComboBox comboBox, ObservableCollection<int> stackedButton)
+        private List<ItemTransferInvoice> fillPdfList(ComboBox comboBox, ObservableCollection<long> stackedButton)
         {
             List<ItemTransferInvoice> list = new List<ItemTransferInvoice>();
             
@@ -497,7 +497,7 @@ namespace Restaurant.View.reports.purchaseReports
             return list;
         }
 
-        private void fillRowChart(ComboBox comboBox, ObservableCollection<int> stackedButton)
+        private void fillRowChart(ComboBox comboBox, ObservableCollection<long> stackedButton)
         {
             MyAxis.Labels = new List<string>();
             List<string> names = new List<string>();
@@ -645,7 +645,7 @@ namespace Restaurant.View.reports.purchaseReports
 
             txt_count.Text = dgInvoice.Items.Count.ToString();
 
-            ObservableCollection<int> selected = new ObservableCollection<int>();
+            ObservableCollection<long> selected = new ObservableCollection<long>();
             if (selectedTab == 0)
                 selected = selectedBranchId;
             if (selectedTab == 1)
