+++ Restaurant/Restaurant/View/reports/salesReports/uc_invoiceSalesReports.xaml.cs
@@ -101,10 +101,10 @@ namespace Restaurant.View.reports.salesReports
         Agent agentModel = new Agent();
         User userModel = new User();
 
-        List<int> selectedBranchId = new List<int>();
-        List<int> selectedPosId = new List<int>();
-        List<int> selectedVendorsId = new List<int>();
-        List<int> selectedUserId = new List<int>();
+        List<long> selectedBranchId = new List<long>();
+        List<long> selectedPosId = new List<long>();
+        List<long> selectedVendorsId = new List<long>();
+        List<long> selectedUserId = new List<long>();
 
         private void UserControl_Loaded(object sender, RoutedEventArgs e)
         {//load
@@ -332,7 +332,7 @@ namespace Restaurant.View.reports.salesReports
             else
                 col_tax.Visibility = Visibility.Visible;
 
-            List<int> selected = new List<int>();
+            List<long> selected = new List<long>();
             if (selectedTab == 0)
                 selected = selectedBranchId;
             if (selectedTab == 1)
@@ -859,7 +859,7 @@ namespace Restaurant.View.reports.salesReports
         #endregion
 
         #region charts
-        private void fillPieChart(List<int> stackedButton)
+        private void fillPieChart(List<long> stackedButton)
         {
             List<string> titles = new List<string>();
             IEnumerable<int> x = null;
@@ -941,7 +941,7 @@ namespace Restaurant.View.reports.salesReports
             chart1.Series = piechartData;
         }
 
-        private void fillColumnChart(List<int> stackedButton)
+        private void fillColumnChart(List<long> stackedButton)
         {
             axcolumn.Labels = new List<string>();
             List<string> names = new List<string>();
@@ -1094,7 +1094,7 @@ namespace Restaurant.View.reports.salesReports
             cartesianChart.Series = columnChartData;
         }
 
-        private void fillRowChart(List<int> stackedButton)
+        private void fillRowChart(List<long> stackedButton)
         {
             MyAxis.Labels = new List<string>();
             List<string> names = new List<string>();
